refactor(*.vue): files architecture
