+++ src/App.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
-  import VerticalNavbar from "./components/VerticalNavbar.vue";
-  import transactions_container from "./components/transactions_container.vue";
-  import ERC from "./components/ERC.vue";
+  import VerticalNavbar from "./components/main/VerticalNavbar.vue";
+  import transactions_container from "./components/transaction/transactions_container.vue";
+  import ERC from "./components/main/ERC/ERC.vue";
 </script>
 
 <template>
