+++ src/components/main/ERC/ERC.vue
@@ -7,13 +7,13 @@ export default defineComponent({
 </script>
 
 <script setup lang="ts">
-import TabsERC from './Tabs.vue';
-import Token from './Token.vue';
-import NFT from './NFT.vue';
+import TabsERC from '@/components/main/ERC/tab/Tabs.vue';
+import Token from './token/Token.vue';
+import NFT from './NFT/NFT.vue';
 import { ref } from 'vue';
-import Actions from './UserActions.vue';
-import UserProfile from './UserProfile.vue'
-import UserInfos from '@/components/UserInfos.vue'
+import Actions from '@/components/main/action-button/UserActions.vue';
+import UserProfile from '../user-info/UserProfile.vue'
+import UserInfos from '@/components/main/user-info/UserInfos.vue'
 
 const currentTab = ref('Token')
 </script>
